+++ src/app/modules/shared/components/copy-to-clipboard/copy-to-clipboard.component.ts
@@ -1,0 +1,18 @@
+import { Component, Input } from '@angular/core';
+import { ElectronService } from '../../services/electron/electron.service';
+
+@Component({
+  selector: 'app-copy-to-clipboard',
+  templateUrl: './copy-to-clipboard.component.html',
+  styleUrls: ['./copy-to-clipboard.component.scss'],
+})
+export class CopyToClipboardComponent {
+  @Input()
+  public toCopy: string;
+
+  constructor(private electron: ElectronService) {}
+
+  public copy(filePath: string) {
+    this.electron.clipboard.writeText(filePath, 'clipboard');
+  }
+}