+++ src/app/modules/translations/services/tree-builder/tree-builder.service.ts
@@ -14,16 +14,18 @@ export class TreeBuilderService {
     let result: ITreeNode[] = [];
     let level = { result };
 
-    items.forEach(path => {
-      path.split(separator).reduce((r, name, idx, parts) => {
-        if (!r[name]) {
-          r[name] = { result: [] };
-          r.result.push({ currentId: name, currentPath: parts.slice(0, idx + 1).join(separator), children: r[name].result });
-        }
+    items
+      .sort((a, b) => a.localeCompare(b))
+      .forEach(path => {
+        path.split(separator).reduce((r, name, idx, parts) => {
+          if (!r[name]) {
+            r[name] = { result: [] };
+            r.result.push({ currentId: name, currentPath: parts.slice(0, idx + 1).join(separator), children: r[name].result });
+          }
 
-        return r[name];
-      }, level);
-    });
+          return r[name];
+        }, level);
+      });
 
     return result;
   }